requirements.txt
biopython==1.70
click==6.7
Flask==0.12.2
Flask-CDN==1.5.3
Flask-WTF==0.14.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.13.3
Werkzeug==0.13
WTForms==2.1
